Rename routes to searchRoutes
