+++ src/features/internal/components/DeeplinksGeneratorForm.tsx
@@ -22,7 +22,7 @@ import {
 import { getProfileStackConfig } from 'features/navigation/ProfileStackNavigator/getProfileStackConfig'
 import { isProfileStackScreen } from 'features/navigation/ProfileStackNavigator/routes'
 import { getScreenPath } from 'features/navigation/RootNavigator/linking/getScreenPath'
-import { isSearchStackScreen } from 'features/navigation/SearchStackNavigator/routes'
+import { isSearchStackScreen } from 'features/navigation/SearchStackNavigator/searchRoutes'
 import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/searchStackHelpers'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { isTabScreen } from 'features/navigation/TabBar/routes'
