+++ src/features/navigation/SearchStackNavigator/__mocks__/searchRoutes.ts
